requirements.txt
# Core FastAPI
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database
motor>=3.3.0
pymongo>=4.6.0

# Data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Authentication
pyjwt>=2.8.0
bcrypt>=4.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0

# HTTP utilities
httpx>=0.25.0
requests>=2.31.0
aiofiles>=23.2.0

# AI Services
google-generativeai>=0.3.0
google-cloud-vision>=3.4.0
pillow>=10.0.0
numpy>=1.24.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.11.0